test: move version-bump spec to main process
